import corrections, for the errors raised by moving the engines module into rsquarelabs-core
