+++ _docs/use_cases.md
@@ -1,1 +1,1 @@
-# User cases 
+# User cases
